fix: refactor initSubAccountConfig
